+++ packages/cli/src/app/index.ts
@@ -1,25 +1,18 @@
-import { sep, join } from 'path'
-import { PackageJson } from 'type-fest'
+import { sep } from 'path'
 import chalk from 'chalk'
 import {
   generator,
   prompt,
   runGenerators,
-  loadJSON,
   fromFile,
   install,
   copyFiles,
   toFile,
   when
 } from '@feathershq/pinion'
-import { FeathersBaseContext, FeathersAppInfo, initializeBaseContext } from '../commons'
-import { generate as authenticationGenerator } from '../authentication'
-import { generate as connectionGenerator } from '../connection'
-
-export type DependencyVersions = { [key: string]: string }
-
-const addVersions = (dependencies: string[], versions: DependencyVersions) =>
-  dependencies.map((dep) => `${dep}@${versions[dep] ? versions[dep] : 'latest'}`)
+import { FeathersBaseContext, FeathersAppInfo, initializeBaseContext, addVersions } from '../commons'
+import { generate as authenticationGenerator, prompts as authenticationPrompts } from '../authentication'
+import { generate as connectionGenerator, prompts as connectionPrompts } from '../connection'
 
 export interface AppGeneratorData extends FeathersAppInfo {
   /**
@@ -42,11 +35,6 @@ export interface AppGeneratorData extends FeathersAppInfo {
    * The source folder where files are put
    */
   lib: string
-  /**
-   * A list dependencies that should be installed with a certain version.
-   * Used for installing development dependencies during testing.
-   */
-  dependencyVersions?: DependencyVersions
 }
 
 export type AppGeneratorContext = FeathersBaseContext &
@@ -59,16 +47,11 @@ export type AppGeneratorArguments = FeathersBaseContext & Partial<AppGeneratorDa
 
 export const generate = (ctx: AppGeneratorArguments) =>
   generator(ctx)
-    .then(
-      loadJSON(join(__dirname, '..', '..', 'package.json'), (pkg: PackageJson) => ({
-        dependencyVersions: {
-          ...pkg.devDependencies,
-          ...ctx.dependencyVersions
-        },
-        dependencies: [],
-        devDependencies: []
-      }))
-    )
+    .then((ctx) => ({
+      ...ctx,
+      dependencies: [],
+      devDependencies: []
+    }))
     .then(
       prompt<AppGeneratorArguments, AppGeneratorContext>((ctx) => [
         {
@@ -124,58 +107,12 @@ export const generate = (ctx: AppGeneratorArguments) =>
             { value: 'yarn', name: 'Yarn' }
           ]
         },
-        {
-          name: 'database',
-          type: 'list',
-          when: !ctx.database,
-          message: 'What is your main database?',
-          suffix: chalk.grey(' Other databases can be added at any time'),
-          choices: [
-            { value: 'mongodb', name: 'MongoDB' },
-            { value: 'knex', name: 'SQL (PostgreSQL, SQLite etc.)', disabled: true }
-          ]
-        },
-        {
-          name: 'connectionString',
-          type: 'input',
-          when: (answers: AppGeneratorArguments) => !ctx.connectionString && answers.database !== 'custom',
-          message: 'Enter your database connection string',
-          default: (answers: AppGeneratorArguments) => `mongodb://localhost:27017/${answers.name}`
-        },
-        {
-          type: 'checkbox',
-          name: 'authStrategies',
-          when: !ctx.authStrategies,
-          message: 'Which user authentication methods do you want to use?',
-          suffix: chalk.grey(' Other methods and providers can be added at any time.'),
-          choices: [
-            {
-              name: 'Email + Password',
-              value: 'local',
-              checked: true
-            },
-            {
-              name: 'Google',
-              value: 'google'
-            },
-            {
-              name: 'Facebook',
-              value: 'facebook'
-            },
-            {
-              name: 'Twitter',
-              value: 'twitter'
-            },
-            {
-              name: 'GitHub',
-              value: 'github'
-            },
-            {
-              name: 'Auth0',
-              value: 'auth0'
-            }
-          ]
-        }
+        ...connectionPrompts(ctx),
+        ...authenticationPrompts({
+          ...ctx,
+          service: 'users',
+          entity: 'user'
+        })
       ])
     )
     .then(runGenerators(__dirname, 'templates'))
@@ -183,7 +120,7 @@ export const generate = (ctx: AppGeneratorArguments) =>
     .then(initializeBaseContext())
     .then(
       when<AppGeneratorContext>(
-        ({ authStrategies, database }) => authStrategies.length > 0 && database !== 'custom',
+        ({ authStrategies }) => authStrategies.length > 0,
         async (ctx) => {
           const { dependencies } = await connectionGenerator(ctx)
 
@@ -242,7 +179,7 @@ export const generate = (ctx: AppGeneratorArguments) =>
     )
     .then(
       install<AppGeneratorContext>(({ language, framework, devDependencies, dependencyVersions }) => {
-        devDependencies.push('nodemon', 'axios', 'mocha')
+        devDependencies.push('nodemon', 'axios', 'mocha', 'cross-env')
 
         if (language === 'ts') {
           devDependencies.push(