+++ packages/cli/src/service/templates/resolver.tpl.ts
@@ -11,13 +11,13 @@ import type {
   ${upperName}Patch,
   ${upperName}Result,
   ${upperName}Query,
-} from '../${schemaPath}'
+} from '${relative}/${schemaPath}'
 import {
   ${camelName}DataSchema,
   ${camelName}PatchSchema,
   ${camelName}ResultSchema,
   ${camelName}QuerySchema
-} from '../${schemaPath}'
+} from '${relative}/${schemaPath}'
 
 
 // Resolver for the basic data model (e.g. creating new entries)
@@ -68,11 +68,11 @@ export const generate = (ctx: ServiceGeneratorContext) =>
   generator(ctx).then(
     renderSource(
       template,
-      toFile(({ lib, folder, kebabName }: ServiceGeneratorContext) => [
+      toFile(({ lib, folder, fileName }: ServiceGeneratorContext) => [
         lib,
         'resolvers',
         ...folder,
-        `${kebabName}.resolver`
+        `${fileName}.resolver`
       ])
     )
   )